fix u64/u128 conversions

+++ crates/hotshot/testing/src/view_generator.rs
@@ -436,7 +436,7 @@ impl TestView {
 
         let block_header = TestBlockHeader {
             block_number: *next_view,
-            timestamp: *next_view,
+            timestamp: *next_view as u128,
             payload_commitment,
             builder_commitment,
             metadata,
